Refactor ClientService to pull pinned clients from Documents
Previously the Pinned documents were retireved by querying a collection; however, this method is not adaptable for per-user dashboards.

Now the pinned clients can be pulled when providing a list of 'pinned' client ids. These 'pinned client ids' will be stored in each individual user's record as a string[] of ids.